requirements.txt
# Pajakia - Indonesian Tax Preparation Platform
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
redis==5.2.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.28.1

# WhatsApp / HTTP
aiohttp==3.11.11

# AI / OCR
openai==1.58.1
pillow==11.1.0

# Document storage
boto3==1.36.2

# Config
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Security / 2FA
pyotp==2.9.0
qrcode[pil]==8.0
cryptography==44.0.0

# CSV/Excel import
openpyxl==3.1.5

# Dev
pytest==8.3.4
pytest-asyncio==0.25.0
